+++ CHANGELOG.md
@@ -45,6 +45,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - message link rendering is now also shown for bot users
 - message logging happens for bot users as well
 - Audit messages now include the user tag
+- Changed the `schedule` command group name to `scheduled`
+- Renamed the `reddit` command to `image` because it just gets an image from reddit.
 
 ## [2.0.1] - 2020-12-20
 ### Changed